Add case for non-existent student delete

+++ 3.00-starting-project/src/main/java/com/luv2code/springmvc/controller/GradebookController.java
@@ -35,6 +35,10 @@ public class GradebookController {
 
 	@GetMapping("/delete/student/{id}")
 	public String deleteStudent(@PathVariable int id, Model m) {
+		if (!studentAndGradeService.checkIfStudentIsNull(id)) {
+			return "error";
+		}
+
 		studentAndGradeService.deleteStudent(id);
 		Iterable<CollegeStudent> collegeStudents = studentAndGradeService.getGradeBook();
 		m.addAttribute("students", collegeStudents);
